+++ crates/miden-objects/src/asset/vault/mod.rs
@@ -60,7 +60,7 @@ impl AssetVault {
     pub fn new(assets: &[Asset]) -> Result<Self, AssetVaultError> {
         Ok(Self {
             asset_tree: Smt::with_entries(
-                assets.iter().map(|asset| (asset.vault_key().as_word(), (*asset).into())),
+                assets.iter().map(|asset| (asset.vault_key().into(), (*asset).into())),
             )
             .map_err(AssetVaultError::DuplicateAsset)?,
         })
@@ -77,7 +77,7 @@ impl AssetVault {
     /// Returns true if the specified non-fungible asset is stored in this vault.
     pub fn has_non_fungible_asset(&self, asset: NonFungibleAsset) -> Result<bool, AssetVaultError> {
         // check if the asset is stored in the vault
-        match self.asset_tree.get_value(&asset.vault_key().as_word()) {
+        match self.asset_tree.get_value(&asset.vault_key().into()) {
             asset if asset == Smt::EMPTY_VALUE => Ok(false),
             _ => Ok(true),
         }
@@ -97,7 +97,7 @@ impl AssetVault {
         match self.asset_tree.get_value(
             &VaultKey::from_account_id(faucet_id)
                 .expect("faucet ID should be of type fungible")
-                .as_word(),
+                .into(),
         ) {
             asset if asset == Smt::EMPTY_VALUE => Ok(0),
             asset => Ok(FungibleAsset::new_unchecked(asset).amount()),
@@ -119,7 +119,7 @@ impl AssetVault {
     ///
     /// The `vault_key` can be obtained with [`Asset::vault_key`].
     pub fn open(&self, vault_key: VaultKey) -> AssetWitness {
-        let smt_proof = self.asset_tree.open(&vault_key.as_word());
+        let smt_proof = self.asset_tree.open(&vault_key.into());
         // SAFETY: The asset vault should only contain valid assets.
         AssetWitness::new_unchecked(smt_proof)
     }
@@ -204,7 +204,7 @@ impl AssetVault {
         asset: FungibleAsset,
     ) -> Result<FungibleAsset, AssetVaultError> {
         // fetch current asset value from the tree and add the new asset to it.
-        let new: FungibleAsset = match self.asset_tree.get_value(&asset.vault_key().as_word()) {
+        let new: FungibleAsset = match self.asset_tree.get_value(&asset.vault_key().into()) {
             current if current == Smt::EMPTY_VALUE => asset,
             current => {
                 let current = FungibleAsset::new_unchecked(current);
@@ -212,7 +212,7 @@ impl AssetVault {
             },
         };
         self.asset_tree
-            .insert(new.vault_key().as_word(), new.into())
+            .insert(new.vault_key().into(), new.into())
             .map_err(AssetVaultError::MaxLeafEntriesExceeded)?;
 
         // return the new asset
@@ -231,7 +231,7 @@ impl AssetVault {
         // add non-fungible asset to the vault
         let old = self
             .asset_tree
-            .insert(asset.vault_key().as_word(), asset.into())
+            .insert(asset.vault_key().into(), asset.into())
             .map_err(AssetVaultError::MaxLeafEntriesExceeded)?;
 
         // if the asset already exists, return an error
@@ -275,7 +275,7 @@ impl AssetVault {
         asset: FungibleAsset,
     ) -> Result<FungibleAsset, AssetVaultError> {
         // fetch the asset from the vault.
-        let new: FungibleAsset = match self.asset_tree.get_value(&asset.vault_key().as_word()) {
+        let new: FungibleAsset = match self.asset_tree.get_value(&asset.vault_key().into()) {
             current if current == Smt::EMPTY_VALUE => {
                 return Err(AssetVaultError::FungibleAssetNotFound(asset));
             },
@@ -291,7 +291,7 @@ impl AssetVault {
             _ => new.into(),
         };
         self.asset_tree
-            .insert(new.vault_key().as_word(), value)
+            .insert(new.vault_key().into(), value)
             .map_err(AssetVaultError::MaxLeafEntriesExceeded)?;
 
         // return the asset that was removed.
@@ -311,7 +311,7 @@ impl AssetVault {
         // remove the asset from the vault.
         let old = self
             .asset_tree
-            .insert(asset.vault_key().as_word(), Smt::EMPTY_VALUE)
+            .insert(asset.vault_key().into(), Smt::EMPTY_VALUE)
             .map_err(AssetVaultError::MaxLeafEntriesExceeded)?;
 
         // return an error if the asset did not exist in the vault.